Refactor RemoveKeywordCollisions->ManipulateNames
Rewrite of RemoveKeywordCollisions into a more generic
ManipulateNames. The new ManipulateNames transform is abstract in a
'(String, Namespace) => String' method that can be used for arbitrary
manipulation of names in a circuit. The software architecture remains
mostly the same (a rename map is used as the underlying data store).
However, the new ManipulateNames used Target as opposed to Named.

Add the ability for naming to be selectively enabled or disabled via:
  - ManipulateNamesWhitelistAnnotation
  - ManipulateNamesBlacklistAnnotation

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>